feat: add native SVG rendering, editing, and undo/redo support
Render SVG thumbnails in the gallery grid and full-resolution in the
viewer through a new Rust pipeline built on resvg and usvg. The pipeline
exposes JNI operations for loading, rendering at arbitrary sizes,
serializing the document model, and applying rotate, flip, and crop
edits in memory.

Editing operations now persist SVG changes reliably with FileChannel.force
and refresh the gallery state through an onImageModified callback. A full
undo/redo stack (Ctrl+Z, Ctrl+Y) tracks edit history for the current
document. Copying an SVG image copies its source text rather than a
rasterized bitmap.

The crop overlay initializes with a sensible default rect instead of
reusing stale state. The viewer handles mouse button 4 and 6 for back
navigation and toggles slideshow on spacebar press. Broken image
placeholders show an icon in both the grid and viewer when decoding
fails.

+++ app/src/main/rust/src/image_pipeline.rs
@@ -254,8 +254,8 @@ fn decode_image(
     let mut output = Vec::with_capacity(16 + pixels.len());
     output.extend_from_slice(&(final_w as i32).to_le_bytes());
     output.extend_from_slice(&(final_h as i32).to_le_bytes());
-    output.extend_from_slice(&(eff_w as i32).to_le_bytes());
-    output.extend_from_slice(&(eff_h as i32).to_le_bytes());
+    output.extend_from_slice(&(eff_w as f32).to_le_bytes());
+    output.extend_from_slice(&(eff_h as f32).to_le_bytes());
     output.extend_from_slice(&pixels);
     Some(output)
 }
